+++ src/test/java/nextstep/subway/favorite/acceptance/FavoriteTestFixture.java
@@ -1,4 +1,4 @@
-package nextstep.subway.favorite;
+package nextstep.subway.favorite.acceptance;
 
 import static org.assertj.core.api.Assertions.assertThat;
